Deploy chat sidebar to Vercel
Extract Express handler logic into shared core.ts so the serverless
functions in api/chat/* and the local dev server stay in sync. Add
vercel.json to bundle the chat-sidebar iframe into public/ and expose
suggest/highlights/health endpoints. Point ChatSidebar's iframe at the
deployed Vercel URL so cBioPortal embeds the hosted app and same-origin
API calls reach the functions.
