+++ .gitignore
@@ -70,3 +70,4 @@ api-e2e/validation.js
 .nx
 .turbo
 .claude/
+.vercel
