+++ packages/chat-sidebar-server/.vercelignore
@@ -1,8 +1,6 @@
 node_modules
-dist
-public
-.vercel
 .env
 .env.local
 .env.*.local
+dist
 *.log
